The ammo works
When you Shoot. It takes away a ammo. Once you run out. The number on the side goes down by one and you can't shoot for 3 seconds. Then it goes back up and the process repeats. Once the number on the right reaches 0 and the number on the left reaches 0. You can't damamge the bad guys at all.